# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: eQuadras API - Gestão e Agendamento Esportivo
  description: "Documentação oficial das APIs REST da plataforma eQuadras para integraç\
    ões de sistemas parceiros, bots e clientes de API. Todos os endpoints incluem\
    \ exemplos reais de requisição e resposta."
  contact:
    name: Suporte Técnico eQuadras
    url: https://github.com/GuilhermeAizzaSano/eQuadras
    email: contato@equadras.app
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: v1.0.0
servers:
- url: https://equadras.app
  description: Servidor de Produção (HTTPS)
- url: /
  description: Servidor Relativo (Mesma Origem)
- url: http://localhost:8080
  description: Ambiente Local (Desenvolvimento)
security:
- BearerAuth: []
- ApiKeyAuth: []
...
paths:
...
  /api/admin/auditoria/estatisticas:
    get:
      tags:
      - Auditoria & Logs - API
      summary: Obter estatísticas de auditoria de hoje
      description: "Papel: ADMIN, exigindo Master Admin. Contadores de hoje: logins,\
        \ falhas de login, ações gerais e cancelamentos."
      operationId: obterEstatisticas
      responses:
        "200":
          description: Contadores de auditoria de hoje
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstatisticasAuditoriaDTO"
              example:
                totalLoginsHoje: 42
                totalFalhasLoginHoje: 3
                totalAcoesHoje: 180
                totalCancelamentosHoje: 2
        "403":
          description: Acesso Negado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/acesso-negado
                title: Acesso Negado
                status: 403
                detail: Acesso restrito ao Administrador Geral do sistema.
                instance: /api/admin/auditoria/estatisticas
                code: ACESSO_NEGADO
                timestamp: 2026-09-29T14:30:00.123456Z
  /api/admin/auditoria:
    get:
      tags:
      - Auditoria & Logs - API
      summary: Listar logs de auditoria com filtros
      description: "Papel: ADMIN, exigindo Master Admin (outro ADMIN recebe 403).\
        \ Página ordenada por `criadoEm` decrescente (ordenação fixa). Todos os filtros\
        \ são opcionais e combináveis."
      operationId: listarLogs
      parameters:
      - name: page
        in: query
        description: "Índice da página, começando em 0."
        required: false
        schema:
          type: string
        example: 0
      - name: size
        in: query
        description: Itens por página. Padrão 10; mínimo 1; máximo 100 (fora da faixa
          é ajustado).
        required: false
        schema:
          type: string
          maximum: 100
          minimum: 1
        example: 10
      - name: usuarioId
        in: query
        description: Filtra pelo ID do usuário que executou a ação.
        required: false
        schema:
          type: integer
          format: int64
        example: 10
      - name: categoria
        in: query
        description: "Filtra pela categoria do evento (enum `CategoriaAuditoria`:\
          \ AUTENTICACAO, AGENDAMENTO, QUADRA, USUARIO, BLOQUEIO, API_KEY)."
        required: false
        schema:
          type: string
          enum:
          - AUTENTICACAO
          - AGENDAMENTO
          - QUADRA
          - USUARIO
          - BLOQUEIO
          - API_KEY
        example: AUTENTICACAO
      - name: acao
        in: query
        description: "Filtra pelo código da ação (ex.: LOGOUT)."
        required: false
        schema:
          type: string
        example: LOGOUT
      - name: dataInicio
        in: query
        description: "Início do período (ISO 8601 com fuso, ex.: `2026-09-29T00:00:00Z`),\
          \ inclusivo."
        required: false
        schema:
          type: string
          format: date-time
        example: 2026-09-29T00:00:00Z
      - name: dataFim
        in: query
        description: Fim do período (ISO 8601 com fuso).
        required: false
        schema:
          type: string
          format: date-time
        example: 2026-09-30T00:00:00Z
      - name: busca
        in: query
        description: Busca livre de texto nos logs.
        required: false
        schema:
          type: string
        example: arthur
      responses:
        "200":
          description: "Página de logs de auditoria, do mais recente ao mais antigo"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageLogAuditoriaResponseDTO"
              example:
                content:
                - id: 981
                  usuarioId: 10
                  usuarioEmail: arthur.prado@email.com
                  usuarioNome: Arthur Prado
                  categoria: AUTENTICACAO
                  acao: LOGOUT
                  entidade: USUARIO
                  recursoId: "10"
                  tipoExecutor: CLIENTE
                  detalhes: Logout efetuado com sucesso.
                  ip: 203.0.113.10
                  userAgent: Mozilla/5.0
                  criadoEm: 2026-09-29T15:40:00Z
                pageable:
                  pageNumber: 0
                  pageSize: 10
                  sort:
                    empty: false
                    sorted: true
                    unsorted: false
                  offset: 0
                  paged: true
                  unpaged: false
                last: true
                totalPages: 1
                totalElements: 1
                size: 10
                number: 0
                sort:
                  empty: false
                  sorted: true
                  unsorted: false
                first: true
                numberOfElements: 1
                empty: false
        "400":
          description: Parâmetro Inválido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/parametro-invalido
                title: Parâmetro Inválido
                status: 400
                detail: "O parâmetro 'categoria' possui um valor inválido: 'X'."
                instance: /api/admin/auditoria
                code: PARAMETRO_INVALIDO
                timestamp: 2026-09-29T14:30:00.123456Z
        "403":
          description: Acesso Negado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/acesso-negado
                title: Acesso Negado
                status: 403
                detail: Acesso restrito ao Administrador Geral do sistema.
                instance: /api/admin/auditoria
                code: ACESSO_NEGADO
                timestamp: 2026-09-29T14:30:00.123456Z
...
components:
  schemas:
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    EstatisticasAuditoriaDTO:
      type: object
      properties:
        totalAcoesHoje:
          type: integer
          format: int64
          description: Total de ações registradas hoje
          example: 180
        totalCancelamentosHoje:
          type: integer
          format: int64
          description: Cancelamentos de reserva hoje
          example: 2
        totalFalhasLoginHoje:
          type: integer
          format: int64
          description: Tentativas de login com falha hoje
          example: 3
        totalLoginsHoje:
          type: integer
          format: int64
          description: Logins bem-sucedidos hoje
          example: 42
    LogAuditoriaResponseDTO:
      type: object
      properties:
        acao:
          type: string
          description: Código da ação executada
          example: LOGOUT
        categoria:
          type: string
          description: "Categoria do evento: AUTENTICACAO, AGENDAMENTO, QUADRA, USUARIO,\
            \ BLOQUEIO ou API_KEY"
          enum:
          - AUTENTICACAO
          - AGENDAMENTO
          - QUADRA
          - USUARIO
          - BLOQUEIO
          - API_KEY
          example: AUTENTICACAO
        criadoEm:
          type: string
          format: date-time
          description: Instante do registro (UTC)
          example: 2026-09-29T15:40:00Z
        detalhes:
          type: string
          description: Detalhes em texto livre
          example: Logout efetuado com sucesso.
        entidade:
          type: string
          description: Entidade afetada
          example: USUARIO
        id:
          type: integer
          format: int64
          description: ID do registro de auditoria
          example: 981
        ip:
          type: string
          description: Endereço IP de origem
          example: 203.0.113.10
        recursoId:
          type: string
          description: ID do recurso afetado
          example: 10
        tipoExecutor:
          type: string
          description: "Tipo do executor da ação (ex.: MASTER_ADMIN, ADMIN_QUADRA,\
            \ CLIENTE)"
          enum:
          - MASTER_ADMIN
          - ADMIN_QUADRA
          - CLIENTE
          - SISTEMA
          example: CLIENTE
        userAgent:
          type: string
          description: User-Agent do cliente
          example: Mozilla/5.0
        usuarioEmail:
          type: string
          description: E-mail do executor no momento da ação
          example: arthur.prado@email.com
        usuarioId:
          type: integer
          format: int64
          description: ID do usuário que executou a ação
          example: 10
        usuarioNome:
          type: string
          description: Nome do executor no momento da ação
          example: Arthur Prado
    PageLogAuditoriaResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LogAuditoriaResponseDTO"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ProblemaErro:
      type: object
      description: "Corpo de erro (RFC 9457). Os filtros de segurança devolvem só\
        \ status, title e detail."
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
        camposIncorretos:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
              mensagem:
                type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      description: "Informe o token JWT ou API-Key no formato: `Bearer eq_...` gerado\
        \ no painel da conta."
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      description: Chave de API pessoal gerada no painel do usuário (formato `eq_...`).
        Válida em todas as rotas da API respeitando os papéis (ROLE_CLIENT e ROLE_ADMIN).
      name: X-API-KEY
      in: header